requirements.txt
Flask
flask-cors
openai>=1.0.0
requests
python-dotenv
gunicorn
pandas
matplotlib
opencv-python
openai-whisper @ git+https://github.com/openai/whisper.git
moviepy==1.0.3
mediapipe
celery[redis] 

